Renamed CourseSelector->CourseSelection for consistency
