+++ src/models/CourseSelection.js
