+++ src/models/index.js
@@ -23,7 +23,7 @@ import Borders from './BordersModel';
 import Flags from './FlagsModel';
 import Projection from './ProjectionModel';
 import Year from './YearModel';
-import CourseSelector from './CourseSelector';
+import CourseSelection from './CourseSelection';
 import Prepared from './PreparedData';
 
 import FeedbackForm from './FeedbackForm';
@@ -48,7 +48,7 @@ export default class storeModel {
     this.projection = new Projection(this);
     this.deck = new DeckViewport(this);
     this.year = new Year(this);
-    this.courseSelection = new CourseSelector(this);
+    this.courseSelection = new CourseSelection(this);
     this.borders = new Borders(this);
     this.properties = new Properties(this);
     this.prepared = new Prepared(this);
